fix: align fastapi tool schemas
- add ToolFactory schema copies + FastAPI tool endpoint split

- add regression tests for nested schemas
